backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
sqlalchemy>=2.0.39
psycopg2-binary>=2.9.11
pydantic>=2.10.0
python-dotenv>=1.1.0
anthropic>=0.86.0
openai>=1.0.0
google-genai>=1.0.0
httpx>=0.27.0
playwright>=1.58.0
